24122022 -fix IoC and DI into classes and i need to solve the logic of search between datetimes

+++ src/main/java/br/com/banco/model/Conta.java
@@ -12,7 +12,7 @@ public class Conta {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id_conta;
+    private Long id_conta;
     private String nome_responsavel;
 
 }
